refactor(backend): move stateless generation into its own router module
Keeps routers/chat.py within the product file line-count ratchet and keeps
the persisting and non-persisting chat surfaces from drifting into shared
state. Behavior is unchanged; the endpoint still reuses execute_chat_stream.
