+++ backend/main.py
@@ -71,6 +71,7 @@ from routers import (
     focus_sessions,
     advice,
     chat_sessions,
+    chat_generation,
     desktop_agent_vm,
     desktop_chat,
     desktop_core,
@@ -207,6 +208,7 @@ app.include_router(staged_tasks.router)
 app.include_router(focus_sessions.router)
 app.include_router(advice.router)
 app.include_router(chat_sessions.router)
+app.include_router(chat_generation.router)
 app.include_router(scores.router)
 app.include_router(tts.router)
 app.include_router(memory_admin.router)
